Split to data.rs
