+++ src/lib.rs
@@ -9,12 +9,17 @@
 //! See the `LICENSE` file for Copyright and license details.
 //!
 
+mod data;
+
+use crate::data::Reading;
+use crate::data::{
+    random_gen_dust_concentration, random_gen_humidity, random_gen_pressure, random_gen_temperature,
+};
 use clap::Parser;
-use rand::{thread_rng, Rng};
 use reqwest::blocking::Client;
 use reqwest::header::CONTENT_TYPE;
-use serde::{Deserialize, Serialize};
-use std::fmt;
+
+use crate::data::AirPurity;
 
 /// Defines the Ambi Mock Client command line interface as a struct
 #[derive(Parser, Debug)]
@@ -29,90 +34,6 @@ pub struct Cli {
     /// Turns verbose console debug output on
     #[clap(short, long)]
     pub debug: bool,
-}
-
-#[derive(Serialize, Deserialize)]
-struct Reading {
-    temperature: f64,
-    humidity: f64,
-    pressure: i32,
-    dust_concentration: f64,
-    air_purity: String,
-}
-
-impl Reading {
-    fn new(
-        temperature: f64,
-        humidity: f64,
-        pressure: i32,
-        dust_concentration: f64,
-        air_purity: String,
-    ) -> Reading {
-        Reading {
-            temperature,
-            humidity,
-            pressure,
-            dust_concentration,
-            air_purity,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-enum AirPurity {
-    Dangerous,
-    High,
-    Low,
-    FreshAir,
-}
-
-impl AirPurity {
-    fn from_value(value: f64) -> AirPurity {
-        match value {
-            value if (f64::MIN..=50.0).contains(&value) => AirPurity::FreshAir,
-            value if value > 50.0 && value <= 100.0 => AirPurity::Low,
-            value if value > 100.0 && value <= 150.0 => AirPurity::High,
-            _ => AirPurity::Dangerous,
-        }
-    }
-}
-
-// implements fmt::Display for AirPurity so that we can call .to_string() on
-// each enum value
-impl fmt::Display for AirPurity {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            AirPurity::Low => write!(f, "Fresh Air"),
-            AirPurity::High => write!(f, "Low Pollution"),
-            AirPurity::Dangerous => write!(f, "High Pollution"),
-            AirPurity::FreshAir => write!(f, "Dangerous Pollution"),
-        }
-    }
-}
-
-fn random_gen_humidity() -> f64 {
-    let mut rng = thread_rng();
-    let value = rng.gen_range(0.0..=100.0);
-    // Limit to 2 decimals
-    f64::trunc(value * 100.0) / 100.0
-}
-
-fn random_gen_temperature() -> f64 {
-    let mut rng = thread_rng();
-    let value = rng.gen_range(15.0..=35.0);
-    // Limit to 2 decimals
-    f64::trunc(value * 100.0) / 100.0
-}
-
-fn random_gen_pressure() -> i32 {
-    let mut rng = thread_rng();
-    rng.gen_range(900..=1100)
-}
-
-fn random_gen_dust_concentration() -> f64 {
-    let mut rng = thread_rng();
-    let value = rng.gen_range(0.0..=1000.0);
-    f64::trunc(value * 100.0) / 100.0
 }
 
 pub fn run(cli: &Cli) {
@@ -165,24 +86,5 @@ pub fn run(cli: &Cli) {
                 }
             }
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn air_purity_from_value_returns_correct_enum() {
-        let mut rng = thread_rng();
-        let fresh_air = rng.gen_range(0.0..=50.0);
-        let low = rng.gen_range(51.0..=100.0);
-        let high = rng.gen_range(101.0..=150.0);
-        let dangerous = rng.gen_range(151.0..f64::MAX);
-
-        assert_eq!(AirPurity::from_value(fresh_air), AirPurity::FreshAir);
-        assert_eq!(AirPurity::from_value(low), AirPurity::Low);
-        assert_eq!(AirPurity::from_value(high), AirPurity::High);
-        assert_eq!(AirPurity::from_value(dangerous), AirPurity::Dangerous);
     }
 }
